Add files via upload
Project description: India has the highest number of road fatalities in the world.
In 2016, for which global figures are available, India accounted for more than a third of global
road accident deaths. The World Health Organization says such deaths are under-reported and
estimated that in 2016, the figure for India was likely twice as big as that reported by the
government.
In this project, perform a detailed statistical data analysis on Road Accidents of India over a
period of 2003 to 2016 years using various Python Libraries.